move the remaining search code out of app.rs
